+++ all-is-cubes/src/math.rs
@@ -6,7 +6,6 @@
 use std::fmt;
 
 use cgmath::{EuclideanSpace as _, Point3, Vector3};
-use noise::NoiseFn;
 use num_traits::identities::Zero;
 pub use ordered_float::{FloatIsNan, NotNan};
 
@@ -110,38 +109,4 @@ pub trait Geometry {
     fn wireframe_points<E>(&self, output: &mut E)
     where
         E: Extend<(Point3<FreeCoordinate>, Option<Rgba>)>;
-}
-
-/// Extension trait for [`noise::NoiseFn`] which makes it usable with our [`GridPoint`]s.
-pub trait NoiseFnExt: NoiseFn<[f64; 3]> {
-    /// Sample the noise at the center of the given cube. That is, convert the integer
-    /// vector to `f64`, add 0.5 to all coordinates, and call [`NoiseFn::get`].
-    ///
-    /// This offset is appropriate for the most resolution-independent sampling, or
-    /// symmetric shapes with even-numbered widths.
-    fn at_cube(&self, cube: GridPoint) -> f64;
-
-    /// As [`NoiseFn::get`], but converting from integer. Unlike [`NoiseFnExt::at_cube`],
-    /// does not apply any offset.
-    fn at_grid(&self, point: GridPoint) -> f64;
-}
-impl<T> NoiseFnExt for T
-where
-    T: NoiseFn<[f64; 3]> + Sized,
-{
-    fn at_cube(&self, cube: GridPoint) -> f64
-    where
-        Self: Sized,
-    {
-        let point = cube.map(f64::from) + Vector3::new(0.5, 0.5, 0.5);
-        NoiseFn::get(&self, point.into())
-    }
-
-    fn at_grid(&self, point: GridPoint) -> f64
-    where
-        Self: Sized,
-    {
-        let point = point.map(f64::from);
-        NoiseFn::get(&self, point.into())
-    }
 }